ima: Enable per-namespace hash table
Provide a per-namespace hash table for storing measurement entries
that have been seen by the namespace.  This fixes the problem of the
current shared hash table, which will prevent the same measurements
from being stored on a namespace's measurement list even if it was
seen by native or other namespaces.

Signed-off-by: Yuqiong Sun <suny@us.ibm.com>