ima: Enable recursive measurements down to the host
Allow measurements to be stored on measurement lists of recursively
down to the native (child->parent->...->native).  Namespace of each
layer independently decide whether or not the measurements will be
stored on its list, based on the namespace's measurement policy.

Changelog:
* Fix ima file flags as S_IRUGO and S_IWUGO and add
CAP_SYS_ADMIN checks for opening ima files in a namespace

Signed-off-by: Yuqiong Sun <suny@us.ibm.com>
Signed-off-by: Mehmet Kayaalp <mkayaalp@linux.vnet.ibm.com>