ima: Add test suite for IMA namespace
userns_child_exec.c: a program that can create a new IMA namespace
by providing clone() syscall with CLONE_NEWIMA flag.

userns_child_exec_key.c: a program that can create a new IMA namespace
with a key placed on the ima keyring of the namespace.  Path of the key
is specified by -k argument.

loadkey.sh: a script that reads /proc/[pid]/ima_keyring to find out
the keyring name for the namespace and the load the key onto that
keyring.  The key is specified by the -k argument in
userns_child_exec_key program.

Changelog:
* Added imaevm.h
* Added check for missing command line parameters

Signed-off-by: Yuqiong Sun <suny@us.ibm.com>
Signed-off-by: Mehmet Kayaalp <mkayaalp@linux.vnet.ibm.com>