ima: measure and appraise kexec image and initramfs
ima_process_measurements() reads and calculates the file's hash, only
if the file is in policy.  When a file is read by the kernel, the file
contents need to be loaded into memory regardless of the policy.

This patch defines a new IMA hook named ima_read_and_process_file() to
load and, at the same time, hash a file.  After reading and hashing the
file, ima_read_and_process_file() calls ima_process_measurement() with
the file hash.  If the file is in policy, an iint is created with the
hash value.

This patch defines two new policy hooks named KEXEC_CHECK and
INITRAMFS_CHECK.  Based on policy, the kexec image and initramfs files
are measured/appraised.

Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>