ima: measure and appraise kexec image
ima_process_measurements() determines whether or not a file is in policy
before calculating the file hash.  Instead of reading the file once for
calculating the file hash and possibly again for loading the file into
memory, this patch defines a new IMA hook named ima_read_file_from_fd()
to read, and at the same time hash, a kexec'ed image.  After reading and
hashing the file, ima_read_file_from_fd() calls ima_process_measurement()
to measure and appraise the file.

This patch defines a new policy "func" named KEXEC_CHECK to measure
and/or appraise the kexec image.

Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>