ima: read firmware only once
Instead of reading the firmware twice, once for measuring/appraising
the firmware and again loading it, this patch reads the firmware
once.  This patch removes ima_fw_from_file() and replaces it with a
new hook named ima_read_file_contents().

As ima_read_file_contents() re-appraises the file each time it is
read, there's no need for the firmware specific cache status.  This
patch removes the firmware specific cache status and replaces it
with the generic read status.

Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>