ima: measure kexec boot command line
In addition to files, other things (eg. boot command line) need to be
measured to attest to the integrity of a running system.  A new IMA
hook named ima_buffer_check() calculates and includes the buffer hash
in the measurement list.  Callers of this hook provide the buffer,
buffer length and a policy identifier.

To measure the boot command line, this patch defines a new policy
identifier named BOOT_CHECK.

Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>